src/process-data/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
aiohttp==3.11.11
pydantic==2.10.4
